Fix onboarding CI failures
The wizard hides workspace chrome during setup, so the factory
soft-delete e2e test could not find the workspace name on the list
page. Assert the name on the settings form instead.

The ESLint budget also grew past the limit. Split the oversized
onboarding files and include the generated-name setter in hook
dependency lists.
