+++ web_src/src/pages/factories/pages/onboarding/OnboardingWireframe.tsx
@@ -12,7 +12,8 @@ import { AnalysisSidePanel } from "./AnalysisSidePanel";
 import { DonePanel } from "./DonePanel";
 import { Shell } from "./OnboardingShell";
 import { WIZARD_STEPS, type IntegrationId, type VcsHostId, type WizardStepId } from "./onboardingFixtures";
-import { IssuesStep, NameStep, RepositoryStep, StartStep, VcsStep } from "./onboardingSteps";
+import { IssuesStep } from "./onboardingIssuesStep";
+import { NameStep, RepositoryStep, StartStep, VcsStep } from "./onboardingSteps";
 import { WizardStepFooter } from "./WizardStepFooter";
 import { useConnectDialog } from "./useConnectDialog";
 import { useOnboardingSetupState, type OnboardingSetupApi } from "./useOnboardingSetupState";
