+++ web_src/src/pages/factories/pages/onboarding/onboardingProvision.ts
@@ -1,0 +1,141 @@
+import type {
+  FactoriesFactory,
+  FactoriesFactoryLine,
+  FactoriesUpdateFactoryOnboardingBody,
+  FactoryLineStep,
+} from "@/api-client";
+import type { IntegrationSelections } from "@/pages/home/InstallIntegrationsSection";
+import { ONBOARDING_EVENT_APPS, ONBOARDING_LINE_APPS } from "@/pages/home/factories";
+import type { InstallFactoryInput } from "@/pages/home/useInstallFactory";
+
+export const DEFAULT_LINE_NAME = "Software delivery";
+
+const PRIMARY_LINE_APP_ENTRYPOINT = ONBOARDING_LINE_APPS[0].entrypointNodeId;
+
+type InstallOnboardingApp = (
+  input: InstallFactoryInput,
+) => Promise<{ canvasId: string; canvasName: string } | undefined>;
+
+export type UpdateOnboarding = (input: FactoriesUpdateFactoryOnboardingBody) => Promise<unknown>;
+
+export interface ProvisionedLine {
+  lineId: string;
+  primaryAppId: string;
+}
+
+// A finished line has one step per bundled app, each calling the app onRun
+// entrypoint. Match on the first entrypoint to recover a line provisioned by an
+// earlier, interrupted attempt.
+function findProvisionedLine(factory: FactoriesFactory | null): FactoriesFactoryLine | undefined {
+  return factory?.lines?.find((line) =>
+    line.steps?.some((step) => step.app?.entrypoint === PRIMARY_LINE_APP_ENTRYPOINT),
+  );
+}
+
+async function installOnboardingApp(args: {
+  factoryId: string;
+  appFactoryId: string;
+  selections: IntegrationSelections;
+  appRepository: string;
+  backlogRepository: string;
+  installFactory: InstallOnboardingApp;
+}): Promise<{ canvasId: string; canvasName: string }> {
+  const installed = await args.installFactory({
+    factoryId: args.appFactoryId,
+    workspaceFactoryId: args.factoryId,
+    integrations: args.selections,
+    installParams: {
+      appRepository: args.appRepository,
+      backlogRepository: args.backlogRepository,
+    },
+    startingTaskPrompt: "",
+    navigateOnComplete: false,
+    startInitialRun: false,
+  });
+  if (!installed?.canvasId) throw new Error(`Failed to create the ${args.appFactoryId} app`);
+  return installed;
+}
+
+// Install each bundled app in order and return the line steps that call them.
+// installFactory clears its pending-canvas ref after each success, so the
+// sequential calls create distinct canvases.
+async function provisionLineApps(args: {
+  factoryId: string;
+  selections: IntegrationSelections;
+  appRepository: string;
+  backlogRepository: string;
+  installFactory: InstallOnboardingApp;
+}): Promise<FactoryLineStep[]> {
+  const steps: FactoryLineStep[] = [];
+  for (const app of ONBOARDING_LINE_APPS) {
+    const installed = await installOnboardingApp({
+      factoryId: args.factoryId,
+      appFactoryId: app.factoryId,
+      selections: args.selections,
+      appRepository: args.appRepository,
+      backlogRepository: args.backlogRepository,
+      installFactory: args.installFactory,
+    });
+    steps.push({
+      type: "runApp",
+      app: { app: installed.canvasId, entrypoint: app.entrypointNodeId },
+    });
+  }
+  return steps;
+}
+
+// Event apps listen for GitHub events and are not factory line steps. Install
+// them even when the line already exists, so a retry after a failed finish
+// still creates PR Closure.
+export async function provisionEventApps(args: {
+  factoryId: string;
+  selections: IntegrationSelections;
+  appRepository: string;
+  backlogRepository: string;
+  installFactory: InstallOnboardingApp;
+}): Promise<void> {
+  for (const appFactoryId of ONBOARDING_EVENT_APPS) {
+    await installOnboardingApp({
+      factoryId: args.factoryId,
+      appFactoryId,
+      selections: args.selections,
+      appRepository: args.appRepository,
+      backlogRepository: args.backlogRepository,
+      installFactory: args.installFactory,
+    });
+  }
+}
+
+export async function provisionLine(args: {
+  factory: FactoriesFactory | null;
+  savedLineId?: string;
+  savedAppId?: string;
+  selections: IntegrationSelections;
+  appRepository: string;
+  backlogRepository: string;
+  installFactory: InstallOnboardingApp;
+  createLine: (input: { name: string; steps: FactoryLineStep[] }) => Promise<FactoriesFactoryLine>;
+  updateOnboarding: UpdateOnboarding;
+}): Promise<ProvisionedLine> {
+  const existing = findProvisionedLine(args.factory);
+  const lineId = args.savedLineId ?? existing?.id;
+  if (lineId) {
+    const primaryAppId = args.savedAppId ?? existing?.steps?.[0]?.app?.app;
+    if (primaryAppId) return { lineId, primaryAppId };
+  }
+
+  const steps = await provisionLineApps({
+    factoryId: args.factory?.id ?? "",
+    selections: args.selections,
+    appRepository: args.appRepository,
+    backlogRepository: args.backlogRepository,
+    installFactory: args.installFactory,
+  });
+  const primaryAppId = steps[0]?.app?.app;
+  if (!primaryAppId) throw new Error("Software delivery apps were not created");
+
+  const line = await args.createLine({ name: DEFAULT_LINE_NAME, steps });
+  if (!line.id) throw new Error("Software delivery line was not created");
+  await args.updateOnboarding({ provisionedAppId: primaryAppId, provisionedLineId: line.id });
+  return { lineId: line.id, primaryAppId };
+}
