+++ web_src/src/pages/factories/pages/onboarding/onboardingStatus.spec.ts
@@ -2,7 +2,13 @@ import { describe, expect, it } from "vitest";
 
 import type { FactoriesFactory } from "@/api-client";
 
-import { initialWizardStep, isFactoryOnboardingComplete } from "./onboardingStatus";
+import {
+  apiIssuesSource,
+  initialOnboardingSelections,
+  initialWizardStep,
+  isFactoryOnboardingComplete,
+  localIssuesSource,
+} from "./onboardingStatus";
 
 describe("isFactoryOnboardingComplete", () => {
   it("returns false while completion time is absent", () => {
@@ -52,5 +58,32 @@ describe("initialWizardStep", () => {
         agentHarness: "AGENT_HARNESS_CLAUDE_CODE",
       }),
     ).toBe("name");
+  });
+});
+
+describe("issue source mapping", () => {
+  it("maps API values to wizard choices", () => {
+    expect(localIssuesSource("ISSUES_SOURCE_VCS")).toBe("vcs");
+    expect(localIssuesSource("ISSUES_SOURCE_SKIP")).toBe("skip");
+    expect(localIssuesSource("ISSUES_SOURCE_UNSPECIFIED")).toBeNull();
+  });
+
+  it("maps wizard choices back to API values", () => {
+    expect(apiIssuesSource("vcs")).toBe("ISSUES_SOURCE_VCS");
+    expect(apiIssuesSource(null)).toBe("ISSUES_SOURCE_UNSPECIFIED");
+  });
+});
+
+describe("initialOnboardingSelections", () => {
+  it("fills github and claude from saved integration IDs", () => {
+    expect(
+      initialOnboardingSelections({
+        vcsIntegrationId: "github-1",
+        agentIntegrationId: "claude-1",
+      }),
+    ).toEqual({
+      github: { id: "github-1", name: "github-1", ready: false },
+      claude: { id: "claude-1", name: "claude-1", ready: false },
+    });
   });
 });
