+++ web_src/src/pages/factories/pages/onboarding/onboardingStatus.ts
@@ -1,6 +1,46 @@
-import type { FactoriesFactory } from "@/api-client";
+import type { FactoriesFactory, FactoriesUpdateFactoryOnboardingBody } from "@/api-client";
+import type { IntegrationSelections } from "@/pages/home/homeIntegrationStatus";
 
-import { WIZARD_STEPS, type WizardStepId } from "./onboardingFixtures";
+import { WIZARD_STEPS, type IssuesChoiceId, type WizardStepId } from "./onboardingFixtures";
+
+export function initialOnboardingSelections(onboarding: FactoriesFactory["onboarding"]): IntegrationSelections {
+  const selections: IntegrationSelections = {};
+  if (onboarding?.vcsIntegrationId) {
+    selections.github = {
+      id: onboarding.vcsIntegrationId,
+      name: onboarding.vcsIntegrationId,
+      ready: false,
+    };
+  }
+  if (onboarding?.agentIntegrationId) {
+    selections.claude = {
+      id: onboarding.agentIntegrationId,
+      name: onboarding.agentIntegrationId,
+      ready: false,
+    };
+  }
+  return selections;
+}
+
+export function localIssuesSource(source?: string): IssuesChoiceId | null {
+  const sources: Record<string, IssuesChoiceId> = {
+    ISSUES_SOURCE_VCS: "vcs",
+    ISSUES_SOURCE_LINEAR: "linear",
+    ISSUES_SOURCE_JIRA: "jira",
+    ISSUES_SOURCE_SKIP: "skip",
+  };
+  return source ? (sources[source] ?? null) : null;
+}
+
+export function apiIssuesSource(source: IssuesChoiceId | null): FactoriesUpdateFactoryOnboardingBody["issuesSource"] {
+  const sources: Partial<Record<IssuesChoiceId, FactoriesUpdateFactoryOnboardingBody["issuesSource"]>> = {
+    vcs: "ISSUES_SOURCE_VCS",
+    linear: "ISSUES_SOURCE_LINEAR",
+    jira: "ISSUES_SOURCE_JIRA",
+    skip: "ISSUES_SOURCE_SKIP",
+  };
+  return source ? sources[source] : "ISSUES_SOURCE_UNSPECIFIED";
+}
 
 export function isFactoryOnboardingComplete(factory: FactoriesFactory | null | undefined): boolean {
   return Boolean(factory?.onboarding?.completedAt);
